Task 7: Implement core flow (7.1–7.6) - Move client logic to Reflector client component; keep page as server component - Implement submit handling, loading state, error feedback, reset - Add rate-limit flag (AI_RATE_LIMIT_ENABLED) on server route - Add 429 auto-retry with countdown and abortable requests - Improve test stability (cleanup in setup)
