+++ .taskmaster/tasks/task_002.txt
@@ -1,6 +1,6 @@
 # Task ID: 2
 # Title: Design System and UI Component Library
-# Status: completed
+# Status: pending
 # Dependencies: 1
 # Priority: high
 # Description: Establish a design system with reusable UI components focusing on accessibility and responsive design.
@@ -59,7 +59,7 @@ Define --container-* variables for different screen sizes (xs: 20rem, sm: 24rem,
 ### Details:
 Create comprehensive documentation covering: breakpoint system usage, mobile-first design approach, container query patterns, component responsiveness examples, touch-friendly interaction guidelines, and testing strategies for different screen sizes.
 
-## 9. Test and Validate Responsive Behavior [done]
+## 9. Test and Validate Responsive Behavior [pending]
 ### Dependencies: None
 ### Description: Thoroughly test responsive layouts and components across all breakpoints to ensure smooth adaptation without overflow or layout issues.
 ### Details:
