+++ .taskmaster/tasks/task_005.txt
@@ -1,6 +1,6 @@
 # Task ID: 5
 # Title: AI Integration Service
-# Status: completed
+# Status: pending
 # Dependencies: 1
 # Priority: high
 # Description: Set up the backend service to handle AI processing of journal entries using OpenAI's API.
