+++ .taskmaster/tasks/tasks.json
@@ -158,7 +158,62 @@
         "priority": "high",
         "dependencies": [4, 5, 6],
         "status": "pending",
-        "subtasks": []
+        "subtasks": [
+          {
+            "id": 1,
+            "title": "Create Main Application Component with State",
+            "description": "Implement a main App component that manages the state between input and reflection display using useState hooks.",
+            "dependencies": [],
+            "details": "Create a new App.jsx component that will serve as the main container. Implement useState hooks for managing the application state including: userInput (string), reflectionResult (object), isLoading (boolean), and error (string). Structure the component to render both the input component and reflection display component. Pass the necessary state and handler functions as props to child components.",
+            "status": "pending",
+            "testStrategy": "Test the initial state values and state transitions. Verify that the component renders both input and reflection components correctly. Check that state is properly passed down to child components."
+          },
+          {
+            "id": 2,
+            "title": "Implement Reflect Now Button Functionality",
+            "description": "Add functionality to the Reflect Now button that processes user input and triggers the API call to generate a reflection.",
+            "dependencies": ["7.1"],
+            "details": "Create a handleSubmit function that validates user input (checking for empty input), sets isLoading to true, and makes the API call to the reflection service. Implement basic error handling to catch and display any API errors. Pass the handleSubmit function to the input component as a prop. Add debouncing to prevent multiple submissions by disabling the button during API calls.",
+            "status": "done",
+            "testStrategy": "Test the submit functionality with valid and invalid inputs. Verify the loading state is properly set during API calls. Test error handling by mocking failed API responses. Ensure the button is disabled during processing."
+          },
+          {
+            "id": 3,
+            "title": "Handle Loading and Error States",
+            "description": "Implement visual feedback for loading states and error handling with user-friendly messages.",
+            "dependencies": ["7.2"],
+            "details": "Create conditional rendering in the App component to show appropriate loading indicators when isLoading is true. Pass the loading state to the reflection display component. Implement error handling that displays user-friendly error messages when API calls fail. Create a simple error message component or use inline messaging that clearly communicates what went wrong and suggests next steps.",
+            "status": "done",
+            "testStrategy": "Test that loading indicators appear and disappear at appropriate times. Verify error messages are displayed correctly for different error scenarios. Test the user experience during loading and error states."
+          },
+          {
+            "id": 4,
+            "title": "Implement Reset Functionality",
+            "description": "Add a clear/reset function that allows users to start a new reflection by clearing all current state.",
+            "dependencies": ["7.3"],
+            "details": "Create a handleReset function that resets all state variables to their initial values (clearing userInput, reflectionResult, error states). Add a Reset or New Reflection button component that calls this function when clicked. Position this button in an intuitive location, such as below the reflection results or in a header area. Ensure the reset function provides a clean slate for starting a new reflection.",
+            "status": "done",
+            "testStrategy": "Test that the reset function properly clears all state variables. Verify the UI returns to its initial state after reset. Test the reset functionality after various application states (after successful reflection, during error state, etc.)."
+          },
+          {
+            "id": 5,
+            "title": "Friendly Error Mapping and Stale Error Clearing",
+            "description": "Provide concise, user-friendly error messages and clear stale errors when the user edits input.",
+            "details": "Implement a function to normalize API errors into human-readable messages (content policy, rate limit, retryable API error, default). Clear error state when the user resumes editing after an error to return to idle. Keep solution minimal and accessible.",
+            "status": "done",
+            "dependencies": [],
+            "parentTaskId": 7
+          },
+          {
+            "id": 6,
+            "title": "Rate-Limit Resilience and Abortable Requests",
+            "description": "Improve resilience against 429s and allow canceling in-flight requests.",
+            "details": "On 429, show retry-after countdown and retry once automatically; disable the button during wait. Wire requests through AbortController and pass signal to fetch so they can be canceled (e.g., on reset).",
+            "status": "done",
+            "dependencies": [],
+            "parentTaskId": 7
+          }
+        ]
       },
       {
         "id": 8,
@@ -284,7 +339,7 @@
     ],
     "metadata": {
       "created": "2025-07-31T04:00:12.936Z",
-      "updated": "2025-07-31T22:21:17.605Z",
+      "updated": "2025-08-08T09:01:38.076Z",
       "description": "Tasks for master context"
     }
   }
