+++ src/lib/ai-client.ts
@@ -24,13 +24,17 @@ export class AIReflectionClient {
   /**
    * Process a journal entry and get AI reflection
    */
-  async processEntry(request: ReflectionRequest): Promise<ReflectionResponse> {
+  async processEntry(
+    request: ReflectionRequest,
+    options?: { signal?: AbortSignal }
+  ): Promise<ReflectionResponse> {
     const response = await fetch(`${this.baseUrl}/api/reflect`, {
       method: 'POST',
       headers: {
         'Content-Type': 'application/json',
       },
       body: JSON.stringify(request),
+      signal: options?.signal,
     })
 
     if (!response.ok) {
